1. i think ai analysis and the results generation in the dasboard are incorrect and in accurate because, when i analyse a project the health score became 0% but after a new file pushed and commited that file alone been analysed as per the flow after analysing that one file total health score became 100% this is large bug which main project itself going wrong and useless makee the ai results more perfect.
2.  now instead going directly to 0% because you give whole number as percentage, use the accurate values of the health score insted 75% it can be 73.6% like that accurate so that it wont go 0% instead 0.4% like that

3.in the dashboard page it is 47.4 but if i click and go inside it is 0% why this misleading informations

+++ DEV--backend/src/main/java/com/developerev/dto/ProjectStatsDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @Builder
 public class ProjectStatsDto {
-    private int healthScore;
+    private double healthScore;
     private int totalFilesAnalyzed;
     private int totalBugs;
     private int totalSecurityIssues;
